sources/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
bcrypt==4.1.2
pyotp==2.9.0
python-dotenv==1.0.0
PyJWT==2.8.0
jwt==1.3.1
windows-curses==2.3.2
mysql-connector==2.2.9
requests==2.31.0
lxml==5.1.0
Google-Images-Search==1.4.6
python-Levenshtein==0.25.0
gTTS==2.5.1
gunicorn==21.2.0
Werkzeug==3.0.1